Change submission_id to BigInt

+++ app/models/submission.py
@@ -3,7 +3,7 @@ from sqlalchemy.dialects.postgresql import JSON
 
 class Submission(db.Model):
     id       = db.Column(db.Integer, primary_key=True)
-    submission_id = db.Column(db.Integer)
+    submission_id = db.Column(db.BigInteger)
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'))
     number_of_points = db.Column(db.Integer)
     revised = db.Column(db.Boolean)
